MOved Tab options in App
