+++ src/Components/Tab.tsx
@@ -1,11 +1,5 @@
 import headlessEngine from "../Engine";
-import {
-  buildTab,
-  TabProps,
-  Tab as HeadlessTab,
-  buildQueryExpression,
-  TabOptions,
-} from "@coveo/headless";
+import { buildTab, TabProps, Tab as HeadlessTab } from "@coveo/headless";
 import React from "react";
 import { Tab as MaterialTab } from "@mui/material";
 export default class Tab extends React.Component<TabProps> {
@@ -37,57 +31,3 @@ export default class Tab extends React.Component<TabProps> {
     );
   }
 }
-
-const filterIntelProcessor = buildQueryExpression()
-  .addStringField({
-    field: "eng_processor",
-    operator: "contains",
-    values: ["Intel"],
-  })
-  .toQuerySyntax();
-
-const filterAmdProcessor = buildQueryExpression()
-  .addStringField({
-    field: "eng_processor",
-    operator: "contains",
-    values: ["AMD"],
-  })
-  .toQuerySyntax();
-
-const noFilter = buildQueryExpression()
-  .addStringField({
-    field: "store_name",
-    operator: "contains",
-    values: ["Barca"],
-  })
-  .toQuerySyntax();
-
-const intelOptions: TabOptions = {
-  id: "Intel",
-  expression: filterIntelProcessor,
-};
-
-const Amd_Options: TabOptions = {
-  id: "AMD",
-  expression: filterAmdProcessor,
-};
-
-const anyOptions: TabOptions = {
-  id: "Any",
-  expression: noFilter,
-};
-
-export const intelProps: TabProps = {
-  initialState: { isActive: false },
-  options: intelOptions,
-};
-
-export const Amd_Props: TabProps = {
-  initialState: { isActive: false },
-  options: Amd_Options,
-};
-
-export const anyProps: TabProps = {
-  initialState: { isActive: true },
-  options: anyOptions,
-};
